requirements.txt
playwright==1.40.0
google-generativeai==0.3.2
python-dotenv==1.0.0
asyncio==3.4.3
